requirements.txt
click==6.7
coverage==4.5.1
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-magic==0.4.15
Wand==0.4.4
Werkzeug==0.14.1
